pyproject.toml
[project]
name = "fast-zero"
version = "0.0.1"
description = ""
authors = [
    {name = "luizsauer",email = "luizsauer@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]>=0.115.14,<0.116.0",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)"
]


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.2"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"

[tool.pytest.ini_options]
pythonpath = "."
addopts = " -p no:warnings"

[tool.ruff]
line-length = 100
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F','E','W','PL','PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest --cov=fast_zero -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
